Quickimmi/DataPipeline/WebCrawlerPipeline/src/WebCrawlerTransformerLambda/requirements.txt
beautifulsoup4==4.12.2
PyPDF2==3.0.1
PyMuPDF==1.23.4
Pillow==9.2.0
html2text==2020.1.16
boto3==1.28.61